попереносив код
