+++ src/js/product-gallery-ex.js
@@ -1,5 +1,6 @@
 import { refs } from './refs';
 import { fetchAllRecipes } from './api-requests';
+import { createMarkupRecipes } from './markup';
 
 async function productGalleryList() {
   try {
@@ -18,56 +19,6 @@ async function productGalleryList() {
   } catch (err) {
     console.log(err);
   }
-}
-
-function createMarkupRecipes(data) {
-  const recipesList = data.results
-    .map(({ _id, title, description, thumb, rating }) => {
-      return `<li class="recipes-item">
-  <div class="photo-thumb">
-    <img class="pic-recipe" src="${thumb}" alt="${title}" loading="lazy" />
-  </div>
-  <button type="button" class="btn-favorite">
-    <svg class="icon-favorite" width="22" height="22">
-      <use href="./img/icons.svg#heart"></use>
-    </svg>
-  </button>
-  <div class="description-recipe-cover">
-    <p class="description-recipe-title">${title}</p>
-    <p class="description-recipe-text">${description}</p>
-    <div class="popular-recept-box"
-      <div class="div-popular-recipe">
-        <div class"star-block"
-        <p class="div-popular-recipe-rating">${rating}</p>
-        <div class="star-recipe-cover-preview">
-        <svg class="star-recipe-reting" width="24" height="24">
-          <use href="/img/icons.svg#star"></use>
-        </svg>
-        <svg class="star-recipe-reting" width="24" height="24">
-          <use href="/img/icons.svg#star"></use>
-        </svg>
-        <svg class="star-recipe-reting" width="24" height="24">
-          <use href="/img/icons.svg#star"></use>
-        </svg>
-        <svg class="star-recipe-reting" width="24" height="24">
-          <use href="/img/icons.svg#star"></use>
-        </svg>
-        <svg class="star-recipe-reting" width="24" height="24">
-        <use href="/img/icons.svg#star"></use>
-        </svg>
-      </div>
-      </div>
-    </div>
- 
-    </div>
-
-    <button class="btn-detail-info" type="button" data-id=${_id}>See recipe</button>
-
-  </div>
-</li>`;
-    })
-    .join('');
-  return recipesList;
 }
 
 /* FULL CARD WITCHOUT CSS
